fix move trainprocess to solve loop (#288)
